+++ src/commands/transactions/index.ts
@@ -12,6 +12,7 @@ export {
 export { updateTransaction, UpdateTransactionOptions } from './update.js';
 export { postTransaction } from './post.js';
 export { checkTransaction } from './check.js';
+export { uncheckTransaction } from './uncheck.js';
 export { trashTransaction } from './trash.js';
 export { untrashTransaction } from './untrash.js';
 export { mergeTransactions } from './merge.js';